continue adding support for fields that are rolled over

+++ src/main/java/org/cowboycoders/ant/profiles/fitnessequipment/pages/GeneralData.java
@@ -113,7 +113,7 @@ public class GeneralData extends CommonPageData implements AntPage, TimeDecodabl
             int temp = seconds.multiply(new BigDecimal(4)).setScale(0, RoundingMode.HALF_UP).intValue();
             RollOverVal val = new RollOverVal(UNSIGNED_INT8_MAX);
             val.setValue(temp);
-            timeElapsed = (int) val.get();
+            timeElapsed = Math.toIntExact(val.get());
             return this;
         }
 
@@ -206,7 +206,7 @@ public class GeneralData extends CommonPageData implements AntPage, TimeDecodabl
             }
             PutUnsignedNumIn1LeBytes(packet, TIME_OFFSET, timeElapsed);
             if (distanceAvailable) {
-                PutUnsignedNumIn1LeBytes(packet, DISTANCE_OFFSET, (int) distance.get());
+                PutUnsignedNumIn1LeBytes(packet, DISTANCE_OFFSET, Math.toIntExact(distance.get()));
             }
             if (speed == null) {
                 PutUnsignedNumIn2LeBytes(packet, SPEED_OFFSET, UNSIGNED_INT16_MAX);
